aiff: fix test suite + cleanups
