+++ aiff/decoder_test.go
@@ -7,7 +7,7 @@ import (
 	"time"
 )
 
-func TestClip(t *testing.T) {
+func TestContainerAttributes(t *testing.T) {
 	expectations := []struct {
 		input           string
 		id              [4]byte
@@ -117,47 +117,6 @@ func Test_Frames(t *testing.T) {
 		if int(clip.Size()) != len(frames) {
 			t.Fatalf("expected %d frames, got %d", clip.Size(), len(frames))
 		}
-	}
-}
-
-func TestClip_Read(t *testing.T) {
-	expectations := []struct {
-		input       string
-		totalFrames int
-	}{
-		{"fixtures/kick.aif", 4484},
-		{"fixtures/delivery.aiff", 17199},
-	}
-
-	for _, exp := range expectations {
-		path, _ := filepath.Abs(exp.input)
-		f, err := os.Open(path)
-		if err != nil {
-			t.Fatal(err)
-		}
-		defer f.Close()
-		d := NewDecoder(f)
-		clip := d.Clip()
-		totalFrames := int(clip.Size())
-		if totalFrames != exp.totalFrames {
-			t.Fatalf("Expected %d frames, got %d\n", exp.totalFrames, totalFrames)
-		}
-		readFrames := 0
-
-		bufSize := 4096
-		buf := make([]byte, bufSize)
-		var n int
-		for readFrames < totalFrames {
-			n, err = clip.Read(buf)
-			if err != nil || n == 0 {
-				break
-			}
-			readFrames += n
-		}
-		if readFrames != totalFrames {
-			t.Fatalf("file expected to have %d frames, only read %d, off by %d frames\n", totalFrames, readFrames, (totalFrames - readFrames))
-		}
-
 	}
 }
